Add files via upload
extracting irregularly gridded data in a netcdf4 into a regular grid by Michael Sumner, with small updates from Pam Michael

the ncdf4 (NCEP Rean, daily) is a sub-ideal test dataset as it is evenly spaced & huge (1 year).  I haven't spent the time getting it to plot the raster as I just used it to update the netcdf4 functions and will try the code on the BLUElink data once I'm on the network.

I need to find an shareable irregularly spaced netcdf4 if including a dataset is kosher.